client/requirements_gpu.txt
absl-py==0.6.1
astor==0.7.1
backcall==0.1.0
bleach==3.0.2
contextlib2==0.5.5
cycler==0.10.0
Cython==0.29
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
gast==0.2.0
greenlet==0.4.15
grpcio==1.16.0
h5py==2.8.0
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.1
mistune==0.8.4
msgpack==0.5.6
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.15.3
object-detection==0.1
opencv-python==3.4.4.19
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
pycocotools==2.0.0
Pygments==2.2.0
pynvim==0.3.1
pyparsing==2.2.2
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-dateutil==2.7.5
pytz==2018.7
pyzmq==17.1.2
qtconsole==4.4.2
Send2Trash==1.5.0
six==1.11.0
tensorboard==1.12.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
